+++ Diagnostics/Assets/Basic/Audiogram/AudiogramController.cs
@@ -15,7 +15,6 @@ using KLib;
 using KLib.Signals;
 using KLib.Signals.Enumerations;
 using KLib.Signals.Waveforms;
-using TMPro.EditorUtilities;
 
 
 public class AudiogramController : MonoBehaviour, IRemoteControllable