+++ Diagnostics/Assets/Turandot/Screen/Turandot.Screen.ScreenElements.cs
@@ -10,12 +10,16 @@ namespace Turandot.Screen
         public List<InputLayout> Inputs { get; set; }
         public bool ApplyCustomScreenColor { get; set; }
         public string finalPrompt = "";
+        public bool ApplyParamSpecificScreenColor { get; set; }
+        public int Color {  get; set; } 
 
         public ScreenElements()
         {
             Cues = new List<CueLayout>();
             Inputs = new List<InputLayout>();
             ApplyCustomScreenColor = false;
+            ApplyParamSpecificScreenColor = false;
+            Color = -1;
         }
     }
 }